node: Use IPv6 node address for CEPs for IPv6 underlays
This commit extends our local node logic to return the IPv6 node address
for CiliumEndpoint objects whenever the underlay is set to IPv6. This is
necessary to have the IPv6 address when populating the tunnel endpoint
in the ipcache. Otherwise, on dual-stack clusters, we would always end
up with the IPv4 address in CiliumEndpoints.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>