bpf:classifiers: add overlay VXLAN/Geneve classifiers
This commits introduces two new classifiers CLS_FLAG_{VXLAN,GENEVE}.
This are used to signal from datapath that the current packet being
traced/dropped is using one of these two protocols as underlay.
This is not used for computing bpf metrics so far, but will be useful
in a subsequent PR for decoding encap'd layer from Hubble/Monitor.
Having different signals for VXLAN and Geneve traffic is actually useful
so that Hubble/Monitor does not need to check the Cilium configuration
to retrieve which tunnel protocol is in use. This way, the datapath
automatically marks every overlay notification, and Hubble can easily
chose between the next decoding layer to pick.

With this patch, we introduce two possible ways of determining a packet
is overlay:

1. ctx->mark:   available only with SKB (XDP does not have marks), we look
                whether the mark contains MARK_MAGIC_OVERLAY. That value
                is usually set on our egress path. This has the advantage
                that is fast, as it does not require to further inspect
                packet L2/L3/L4 headers.
2. raw headers: available both with XDP and SKB (yes, we can still have
                trace notifications in XDP), this method inspects the
                whole packet layers up to L4 UDP. By looking at protocol
                and known ports (TUNNEL_PORT in use), we can determine
                a packet is overlay.

While (2) can always work in every condition, (1) works only when the mark
is set. This means that in all codepaths where the mark is cleared it can
fail. It would be still backed up by logic in (2), given that it will
always run in case (1) was not informative enough.

As of today, I'm not aware of codepaths where while egressing we clear the
OVERLAY mark, given we rely on that (ex. in cil_to_netdev, for
wireguard/ipsec encryption, etc.). If this will turn out not to be solid
enough, we can patch such trace notifications to accept an additional
parameter `__u32 magic`, which is the magic value computed at the
beginning of the functions, before clearing the mark and invoking the trace event.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>