bpf:optimization: distinguish trace paths IPv4/IPv6 when known
This commit introduces back the `send_trace_notify{4,6}` helpers to be used
in known IPv4/IPv6 codepaths. This can help with out packet classification
logic: we now rely on the whole L2/L3 protocol parsing to inspect whether
a packet is Overlay. However, in most of the codepaths where we invoke such
trace functions, we already know the protocol type (i.e. IPv4 or IPv6 path).
In that case, we would not need to parse the whole packet and protocol from the beginning.

This patch changes `send_trace_notify` to `send_trace_notify{4,6}` when
known at the time of calling such helpers. In most of the modified files,
this is just a matter of patching the helper name. In the following, a few
more changes were introduced:

1. cil_from_container: let's try to call a more informative trace helper
   depending on the taken codepath. This requires a bit of code duplication,
   but surely helps lowering complexity in the IPv4/IPv6 case.
2. cil_lxc_policy_egress: same as above.
3. cil_to_container: same as above.
4. cil_from_overlay: same above.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>