bpf:classifiers: accept protocol when known in IPv4/IPv6 paths
This commits patches the current `ctx_classify()` logic to accept an
additional parameter `__be16 proto` that is the layer 3 protocol.
This helps skipping the whole L2 parsing logic in case the protocol is
already known when invoking the function: for instance, from
`send_trace_notify4()` we pass `bpf_ntohs(ETH_P_IP)`, and similarly
`bpf_ntohs(ETH_P_IPV6)` for the v6 path. When unknown, such as from
`send_drop_notify()` or the generic `send_trace_notify()`, this is equal
to 0, and the `ctx_classify()` function will retrieve the protocol itself
from the packet L2 header, if needed.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>