bpf/tests: Rewrite ipsec_redirect test for better coverage
This commit is a rewrite of the ipsec_redirect BPF tests introduced in
commit 33e383f66266 ("ipsec: update unit tests for VinE support"). The
second test in the same file, ipsec_redirect_bad_identities, is kept as
is. The rewrite keeps the same intent but extends the scope.

- The test was a unit test more than an integration test, meaning it
  only covered function ipsec_maybe_redirect_to_encrypt without covering
  how that function is integrated in our datapath.
- As a result, the setup of the test had been mixed with the actual
  checks. That is, the test didn't have a proper SETUP() block.

This commit extends the test coverage to cover the whole cil_from_netdev
BPF program in bpf_host instead of just ipsec_maybe_redirect_to_encrypt.

Fixes: 33e383f66266 ("ipsec: update unit tests for VinE support")
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>