bpf/tests: Fix L2 address check in ipsec_redirect
In ipsec_redirect, the destination MAC address rewriting was never
actually checked because of a typo in the test (single equal sign in
assert). As a result, the assert is always passing. After fixing that,
the test started failing because the destination MAC address is
rewritten to cilium_net's MAC address and not to the source MAC address
as encoded in the test.

Fixes: 33e383f66266 ("ipsec: update unit tests for VinE support")
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>